Major code refactor and added some component
