+++ api/src/schema.graphql
@@ -7,7 +7,7 @@ directive @id(
 ) on FIELD_DEFINITION
 
 type Company {
-  comapnyId: ID! @id(autogenerate: true)
+  id: ID! @id(autogenerate: true)
   name: String
   description: String
   website: String
@@ -31,7 +31,7 @@ type Company {
 }
 
 type Domain {
-  domainId: ID! @id(autogenerate: true)
+  id: ID! @id(autogenerate: true)
   name: String
   description: String
   childDomains: [Domain] @relationship(type: "INCLUDES", direction: OUT)
@@ -39,7 +39,7 @@ type Domain {
 }
 
 type Offering {
-  offeringId: ID! @id(autogenerate: true)
+  id: ID! @id(autogenerate: true)
   name: String
   description: String
   primaryDomain: [Domain] @relationship(type: "PRIMAY_DOMAIN", direction: OUT)
