+++ web-react/src/components/Domain/CreateDomain.css
@@ -1,14 +1,5 @@
-.root {
-  max-width: 700;
-  margin-top: 3px;
-  overflow-x: none;
-  margin: auto;
-}
-
-.textField {
-  margin-left: 1;
-  margin-right: 1;
-  min-width: 300;
+.createDomain {
+  width: 50%;
 }
 
 .hidden-message {
@@ -20,11 +11,6 @@
   color: green;
 }
 
-.navLink {
-  text-decoration: none;
-  color: inherit;
-}
-
-.submitButton {
-  margin: 2px;
+.MuiButtonBase-root.submitButton {
+  margin: 0.3rem;
 }
