+++ web-react/src/components/Domain/CreateDomain.js
@@ -7,38 +7,27 @@ import {
     Button,
     Typography,
 } from '@mui/material'
-import Title from '../Title'
-import { Link } from 'react-router-dom'
-
-
-
+import Heading from '../Heading'
 
 const CREATE_DOMAIN = gql`
  mutation userCreateMutationQuery($input: [DomainCreateInput!]!){
  createDomains(input: $input) {
     domains {
-      domainId
+      id
       name
       description
     }
   }
  }
- 
- 
 `
 function CreateDomain() {
 
-    const [domainId, setDoaminId] = React.useState("")
+    /*  const [domainId, setDoaminId] = React.useState("") */
     const [domainName, setDomainName] = React.useState("")
     const [domainDescription, setDomainDescription] = React.useState("")
 
-
     const [createDomains, { data: mutationData, loading: mutationLoading, error: mutationError }] = useMutation(CREATE_DOMAIN,
         { variables: { input: { name: domainName, description: domainDescription } } })
-
-
-
-
 
     const onDomainNameChange = (e) => {
         const domainName = e.target.value
@@ -50,15 +39,17 @@ function CreateDomain() {
         setDomainDescription(domainDes)
     }
 
+
     const handlerSubmit = (e) => {
         e.preventDefault()
         createDomains()
+        console.log(mutationData)
         console.log(mutationError)
         console.log(mutationLoading)
         if (!mutationError) {
             setDomainName("")
             setDomainDescription("")
-            setDoaminId(mutationData)
+
         }
 
     }
@@ -66,20 +57,10 @@ function CreateDomain() {
 
 
     return (
-        <Paper className="root">
-            <div className='title-container'>
-                <Title>
-                    Create Domain
-                </Title>
-                <Link to="/domain" className="navLink"> <Button color="primary" variant="outlined" >
-                    Domain List
-                </Button></Link>
-            </div>
-
-            <p className={domainId ? 'Success-message' : 'hidden-message'}>{`Domain with domainId ${domainId} is created`}</p>
+        <Paper className="root createDomain">
+            <Heading title=" Create Domain" linkName="Domain List"></Heading>
 
             <form onSubmit={handlerSubmit}>
-
                 <Typography>
                     <TextField className="textField" required label="Domain Name" onChange={onDomainNameChange} value={domainName}>
                     </TextField>
@@ -88,11 +69,15 @@ function CreateDomain() {
                     <TextField className="textField" required label="Domain Description" onChange={onDomainDesChange} value={domainDescription}>
                     </TextField>
                 </Typography>
-
-                <Button className="submitButton" type='submit' color='primary'>
-                    Create Domain
-                </Button>
+                <div className='button-container'>
+                    <Button className="formButton" type='submit' variant="contained" color='primary'>
+                        Create Domain
+                    </Button>
+                </div>
             </form>
+            {/* <div>
+                <p className={domainId ? 'Success-message' : 'hidden-message'}>{`Domain with domainId ${domainId} is created`}</p>
+            </div> */}
         </Paper>
     )
 }
