+++ web-react/src/components/Domain/DomainList.css
@@ -1,6 +1,6 @@
 .MuiFormControl-root.TextField {
   margin: 0.5rem 0;
-  width: 100%;
+  width: 650px;
 }
 
 .labelField {
@@ -8,12 +8,6 @@
   margin: 0.5rem;
 }
 
-.title-container {
-  display: flex;
-  justify-content: space-between;
-  margin: 1rem;
-  margin-bottom: 0;
-}
 
 .textfield-container {
   display: flex;
@@ -26,5 +20,10 @@
 }
 
 .MuiButtonBase-root.formButton {
-  margin: 0.5rem 0 0.5rem 0.7rem;
+  margin: 0.5rem 0.3rem 0.5rem 0.3rem;
+}
+
+.formSelect {
+  width: 100%;
+  margin: 0.5rem 0;
 }
